fix: 複製、コピー、消去操作をそれぞれ異なるコンポーネントに切り出す
