+++ app/_components/item-register/item-register-container.tsx
@@ -1,12 +1,6 @@
 "use client";
 import { TFormOptionIndexResponse } from "@/app/_api/item_register/form/getFormOptionIndex";
-import postItemsDuplicate from "@/app/_api/item_register/postitemsDuplicate";
-import fetchPreregisteredDataShow, {
-  TPreregisteredDataResponse,
-} from "@/app/_api/item_register/preregistered_data/fetchPreregisteredDataShow";
-import fetchPreregisteredDataCountRegistered, {
-  TPreregisteredDataCountRegisteredResponse,
-} from "@/app/_api/item_register/preregistered_data/fetchePreregisteredDataCountRegistered";
+import { TPreregisteredDataResponse } from "@/app/_api/item_register/preregistered_data/fetchPreregisteredDataShow";
 import {
   TItemsShowResponse,
   TMeasurement,
@@ -25,27 +19,17 @@ import {
   getPatternName,
   getSubColorName,
 } from "@/app/_utils/functions/getItemOptionName";
-import QrCode2Icon from "@mui/icons-material/QrCode2";
-import {
-  Box,
-  Button,
-  Dialog,
-  DialogActions,
-  DialogContent,
-  DialogTitle,
-  Fab,
-  SelectChangeEvent,
-} from "@mui/material";
-import { ChangeEvent, useReducer, useState } from "react";
+import { Box } from "@mui/material";
+import { useReducer, useState } from "react";
 import Header from "../common/pages/header";
+import ItemCopyDialog from "./item/item-copy-dialog";
 import ItemDuplicationDialog from "./item/item-duplication-dialog";
 import SizeMeasurementSwitcher from "./item/item-info/size-measurement-switcher";
-import ItemOperationDialog from "./item/item-list/item-operation-dialog";
 import RegisteredItemCardList, {
   TCardInfo,
 } from "./item/item-list/registered-item-card-list";
-import SizeSelectionDialog from "./item/size-selection-dialog";
-import StockingIdInputDialog from "./item/stocking-id-input-dialog";
+import ItemOperationDialog from "./item/item-operation-dialog";
+import PreregisteredItemFetcher from "./item/preregistered-item/preregistered-item-fetcher";
 
 export type TSize = "S" | "L" | "M" | "XL" | "";
 
@@ -91,37 +75,18 @@ type TProps = {
 };
 
 export default function ItemRegisterContainer({ formOption }: TProps) {
-  // 登録コード追加ダイアログ関連
-  const [isStockingIdInputDialogOpen, setIsStockingIdInputDialogOpen] =
-    useState<boolean>(false);
-  const [stockingDateCode, setStockingDateCode] = useState<string>();
-  const [stockingSequence, setStockingSequence] = useState<string>();
   // カード関連
   const [selectedCardId, setSelectedCardId] = useState<number>();
-
-  // サイズ選択ダイアログ関連
-  const [sizeSelectionState, setSizeSelectionState] =
-    useState<TPreregisteredDataCountRegisteredResponse>();
-
-  // preregisteredData関連
-  const [preregisteredData, setPreregisteredData] =
-    useState<TPreregisteredDataResponse>();
 
   // 複製ダイアログ関連
   const [duplicationTargetItemId, setDuplicationTargetItemId] =
     useState<number>();
-  const [selectedCreateNum, setSelectedCreateNum] = useState<
-    number | undefined
-  >(1);
-  const [selectedAdminId, setSelectedAdminId] = useState<number>(2);
+
+  // カード操作ダイアログ関連
+  const [operationTargetCardId, setOperationTargetCardId] = useState<number>();
 
   // カードコピーダイアログ関連
   const [copyTargetCardId, setCopyTargetCardId] = useState<number>();
-
-  // カード消去ダイアログ関連
-  const [operationTargetCardId, setOperationTargetCardId] = useState<number>();
-  const [isDeleteConfirmDialogOpen, setIsDeleteConfirmDialogOpen] =
-    useState<boolean>(false);
 
   const cardsStateReducer = (
     cardsState: TCardsState[],
@@ -164,153 +129,57 @@ export default function ItemRegisterContainer({ formOption }: TProps) {
     });
   };
 
-  const handleClickSize = (selectedSize: TSize) => {
-    // copyTargetCardIdがある場合はコピー
-    if (copyTargetCardId !== undefined || operationTargetCardId !== undefined) {
-      const copyTargetCardState = cardsState.find((cardState, index) => {
-        if (index === copyTargetCardId || index === operationTargetCardId)
-          return cardState;
-      });
-      if (copyTargetCardState) {
-        dispatch({
-          type: "ADD_CARD",
-          value: [
-            {
-              ...copyTargetCardState,
-              isRegistered: false,
-              originalSize: selectedSize,
-              size: selectedSize,
-            },
-          ],
-        });
-      }
-      setCopyTargetCardId(undefined);
-      setOperationTargetCardId(undefined);
-    } else {
-      if (preregisteredData) {
-        const haveAlreadyRegisteredItem = (): number | undefined => {
-          switch (selectedSize) {
-            case "S":
-              if (preregisteredData.smallRefItemId !== null) {
-                return preregisteredData.smallRefItemId;
-              }
-              return undefined;
-            case "M":
-              if (preregisteredData.mediumRefItemId !== null) {
-                return preregisteredData.mediumRefItemId;
-              }
-              return undefined;
-            case "L":
-              if (preregisteredData.largeRefItemId !== null) {
-                return preregisteredData.largeRefItemId;
-              }
-              return undefined;
-            case "XL":
-              if (preregisteredData.extraLargeRefItemId !== null) {
-                return preregisteredData.extraLargeRefItemId;
-              }
-              return undefined;
-            default:
-              return undefined;
-          }
-        };
-        if (haveAlreadyRegisteredItem() !== undefined) {
-          setDuplicationTargetItemId(haveAlreadyRegisteredItem());
-        } else {
-          dispatch({
-            type: "ADD_CARD",
-            value: [
-              {
-                stockingOrderId: preregisteredData.tStockingOrderId,
-                itemImage: preregisteredData.tPreregisteredItem.itemImageUrl,
-                cateSmall: getCateSmallName(
-                  formOption.categorySmalls,
-                  preregisteredData.tPreregisteredItem.mCateSmallId
-                ),
-                brand: getBrandName(
-                  formOption.brands,
-                  preregisteredData.tPreregisteredItem.mBrandId
-                ),
-                color: getColorName(
-                  formOption.colors,
-                  preregisteredData.tPreregisteredItem.mColorId
-                ),
-                subColor: getSubColorName(
-                  formOption.colors,
-                  preregisteredData.tPreregisteredItem.mSubColorId
-                ),
-                pattern: getPatternName(
-                  formOption.patterns,
-                  preregisteredData.tPreregisteredItem.mPatternId
-                ),
-                logo: getLogoName(
-                  formOption.logos,
-                  preregisteredData.tPreregisteredItem.mLogoId
-                ),
-                originalSize: selectedSize,
-                dropSize:
-                  DROP_SIZE[preregisteredData.tPreregisteredItem.dropSize],
-                size: undefined,
-                adminId: undefined,
-                isRegistered: false,
-                shoulder: null,
-                bust: null,
-                waist: null,
-                minWaist: null,
-                maxWaist: null,
-                lengthTop: null,
-                roundNeck: null,
-                hip: null,
-                roundLeg: null,
-                outseam: null,
-                sleeveLength: null,
-                hemWidth: null,
-              },
-            ],
-          });
-          setPreregisteredData(undefined);
-        }
-      }
-    }
-
-    setSizeSelectionState(undefined);
-  };
-
-  const handleClickAdd = () => {
-    if (stockingDateCode !== undefined && stockingSequence !== undefined) {
-      fetchPreregisteredDataShow({
-        id: stockingDateCode + "-" + stockingSequence,
-      })
-        .then((preregisteredDataResponse: TPreregisteredDataResponse) => {
-          setPreregisteredData(preregisteredDataResponse);
-          fetchPreregisteredDataCountRegistered({
-            id: preregisteredDataResponse.tStockingOrderId,
-          })
-            .then(
-              (
-                countDataResponse: TPreregisteredDataCountRegisteredResponse
-              ) => {
-                setSizeSelectionState(countDataResponse);
-              }
-            )
-            .catch((e) => {
-              alert(
-                `データ取得に失敗しました。 ${
-                  (e.response?.data as { message: string })?.message
-                }`
-              );
-            });
-        })
-        .catch((e) => {
-          alert(
-            `データ取得に失敗しました。 ${
-              (e.response?.data as { message: string })?.message
-            }`
-          );
-        });
-
-      setIsStockingIdInputDialogOpen(false);
-    }
+  const createNewCard = (
+    data: TPreregisteredDataResponse,
+    selectedSize: TSize
+  ) => {
+    dispatch({
+      type: "ADD_CARD",
+      value: [
+        {
+          stockingOrderId: data.tStockingOrderId,
+          itemImage: data.tPreregisteredItem.itemImageUrl,
+          cateSmall: getCateSmallName(
+            formOption.categorySmalls,
+            data.tPreregisteredItem.mCateSmallId
+          ),
+          brand: getBrandName(
+            formOption.brands,
+            data.tPreregisteredItem.mBrandId
+          ),
+          color: getColorName(
+            formOption.colors,
+            data.tPreregisteredItem.mColorId
+          ),
+          subColor: getSubColorName(
+            formOption.colors,
+            data.tPreregisteredItem.mSubColorId
+          ),
+          pattern: getPatternName(
+            formOption.patterns,
+            data.tPreregisteredItem.mPatternId
+          ),
+          logo: getLogoName(formOption.logos, data.tPreregisteredItem.mLogoId),
+          originalSize: selectedSize,
+          dropSize: DROP_SIZE[data.tPreregisteredItem.dropSize],
+          size: undefined,
+          adminId: undefined,
+          isRegistered: false,
+          shoulder: null,
+          bust: null,
+          waist: null,
+          minWaist: null,
+          maxWaist: null,
+          lengthTop: null,
+          roundNeck: null,
+          hip: null,
+          roundLeg: null,
+          outseam: null,
+          sleeveLength: null,
+          hemWidth: null,
+        },
+      ],
+    });
   };
 
   const cardInfo: TCardInfo[] = cardsState.map((card, index) => {
@@ -364,164 +233,96 @@ export default function ItemRegisterContainer({ formOption }: TProps) {
     }
   );
 
-  const handleClickDuplicate = () => {
-    if (duplicationTargetItemId && selectedCreateNum)
-      postItemsDuplicate({
-        id: duplicationTargetItemId,
-        tAdminId: selectedAdminId,
-        createNum: selectedCreateNum,
-      })
-        .then((res: TItemsShowResponse[]) => {
-          dispatch({
-            type: "ADD_CARD",
-            value: res.map((item) => {
-              return {
-                ...item,
-                stockingOrderId: item.tStockingOrderId,
-                itemImage: item.itemImageUrl,
-                itemId: item.id,
-                cateSmall: item.mCateSmall.name,
-                brand: item.mBrand.name,
-                color: item.mColor.name,
-                subColor: item.mSubColor?.name ?? "無し",
-                pattern: item.mPattern.name,
-                logo: item.mLogo.name,
-                originalSize: ORIGINAL_SIZE_LABEL[
-                  item.originalSize as keyof typeof ORIGINAL_SIZE_LABEL
-                ] as TSize,
-                dropSize: item.dropSize.name,
-                size: item.size,
-                adminId: item.tAdmin.id,
-                isRegistered: true,
-              };
-            }),
-          });
-        })
-        .catch((e) => {
-          alert(
-            `アイテムの複製に失敗しました。 ${
-              (e.response?.data as { message: string })?.message
-            }`
-          );
-        });
-    setPreregisteredData(undefined);
-    setDuplicationTargetItemId(undefined);
+  const duplicateCard = (data: TItemsShowResponse[]) => {
+    dispatch({
+      type: "ADD_CARD",
+      value: data.map((item: TItemsShowResponse) => {
+        return {
+          ...item,
+          stockingOrderId: item.tStockingOrderId,
+          itemImage: item.itemImageUrl,
+          itemId: item.id,
+          cateSmall: item.mCateSmall.name,
+          brand: item.mBrand.name,
+          color: item.mColor.name,
+          subColor: item.mSubColor?.name ?? "無し",
+          pattern: item.mPattern.name,
+          logo: item.mLogo.name,
+          originalSize: ORIGINAL_SIZE_LABEL[
+            item.originalSize as keyof typeof ORIGINAL_SIZE_LABEL
+          ] as TSize,
+          dropSize: item.dropSize.name,
+          size: item.size,
+          adminId: item.tAdmin.id,
+          isRegistered: true,
+        };
+      }),
+    });
   };
 
-  const handleClickCopy = (copyCardId: number) => {
-    fetchPreregisteredDataCountRegistered({
-      id: cardsState[copyCardId].stockingOrderId,
-    })
-      .then((countDataResponse: TPreregisteredDataCountRegisteredResponse) => {
-        setSizeSelectionState(countDataResponse);
-      })
-      .catch((e) => {
-        alert(
-          `データ取得に失敗しました。 ${
-            (e.response?.data as { message: string })?.message
-          }`
-        );
-      });
-  };
-
-  const deleteCard = () => {
-    if (operationTargetCardId !== undefined)
-      dispatch({
-        type: "DELETE_CARD",
-        value: operationTargetCardId,
-      });
+  const deleteCard = (cardId: number) => {
+    dispatch({
+      type: "DELETE_CARD",
+      value: cardId,
+    });
     setOperationTargetCardId(undefined);
-    setIsDeleteConfirmDialogOpen(false);
   };
 
-  const deleteTargetItemId =
-    operationTargetCardId !== undefined
-      ? cardsState[operationTargetCardId].itemId
-      : undefined;
+  const copyCard = (cardId: number, selectedSize: TSize) => {
+    const copyTargetCardState = cardsState.find((cardState, index) => {
+      if (index === cardId) return cardState;
+    });
+
+    if (copyTargetCardState)
+      dispatch({
+        type: "ADD_CARD",
+        value: [
+          {
+            ...copyTargetCardState,
+            isRegistered: false,
+            originalSize: selectedSize,
+            size: selectedSize,
+          },
+        ],
+      });
+  };
 
   return (
     <>
       {operationTargetCardId !== undefined && (
         <ItemOperationDialog
-          deleteTargetItemId={deleteTargetItemId}
-          isDeleteConfirmDialogOpen={isDeleteConfirmDialogOpen}
-          sizeSelectionState={sizeSelectionState}
+          cardId={operationTargetCardId}
+          cardsState={cardsState}
+          onCloseDialog={() => setOperationTargetCardId(undefined)}
           deleteCard={deleteCard}
-          onClickDelete={() => setIsDeleteConfirmDialogOpen(true)}
-          onClickCancelOperationDialog={() =>
-            setOperationTargetCardId(undefined)
-          }
-          onClickCopy={() => {
-            handleClickCopy(operationTargetCardId);
-          }}
-          onClickCancelConfirmDialog={() => {
-            setOperationTargetCardId(undefined);
-            setIsDeleteConfirmDialogOpen(false);
-          }}
+          copyCard={copyCard}
         />
       )}
-      {copyTargetCardId !== undefined && !sizeSelectionState && (
-        <Dialog open>
-          <DialogTitle>コピー</DialogTitle>
-          <DialogContent>続けて同アイテムを登録しますか?</DialogContent>
-          <DialogActions>
-            <Button onClick={() => setCopyTargetCardId(undefined)}>
-              キャンセル
-            </Button>
-            <Button onClick={() => handleClickCopy(copyTargetCardId)}>
-              OK
-            </Button>
-          </DialogActions>
-        </Dialog>
+
+      {copyTargetCardId !== undefined && (
+        <ItemCopyDialog
+          cardId={copyTargetCardId}
+          onCloseDialog={() => setCopyTargetCardId(undefined)}
+          cardsState={cardsState}
+          copyCard={copyCard}
+        />
       )}
       {duplicationTargetItemId && (
         <ItemDuplicationDialog
-          adminOption={formOption.admins}
-          selectedCreateNum={selectedCreateNum}
-          selectedAdminId={selectedAdminId}
-          onChangeSelectedCreateNum={(
-            e: ChangeEvent<HTMLInputElement | HTMLTextAreaElement>
-          ) =>
-            setSelectedCreateNum(
-              isNaN(parseInt(e.target.value))
-                ? undefined
-                : parseInt(e.target.value)
-            )
-          }
-          onChangeAdminId={(e: SelectChangeEvent<number>) =>
-            setSelectedAdminId(Number(e.target.value))
-          }
-          onClose={() => {
+          cardId={duplicationTargetItemId}
+          onCloseDialog={() => {
             setDuplicationTargetItemId(undefined);
-            setPreregisteredData(undefined);
           }}
-          onClickAdd={handleClickDuplicate}
+          duplicateCard={duplicateCard}
+          adminOption={formOption.admins}
         />
       )}
-      {sizeSelectionState ? (
-        <SizeSelectionDialog
-          stockingId={sizeSelectionState.stockingId}
-          arrivedSmallNum={sizeSelectionState.arrivedSmallNum}
-          arrivedMediumNum={sizeSelectionState.arrivedMediumNum}
-          arrivedLargeNum={sizeSelectionState.arrivedLargeNum}
-          arrivedExtraLargeNum={sizeSelectionState.arrivedExtraLargeNum}
-          registeredSmallNum={sizeSelectionState.registeredSmallNum}
-          registeredMediumNum={sizeSelectionState.registeredMediumNum}
-          registeredLargeNum={sizeSelectionState.registeredLargeNum}
-          registeredExtraLargeNum={sizeSelectionState.registeredExtraLargeNum}
-          onClickSize={handleClickSize}
-          onClose={() => {
-            setSizeSelectionState(undefined);
-            setPreregisteredData(undefined);
-            setOperationTargetCardId(undefined);
-          }}
-        />
-      ) : selectedCardId !== undefined && selectedCardState !== undefined ? (
+      {selectedCardId !== undefined && selectedCardState !== undefined ? (
         <SizeMeasurementSwitcher
           cardState={selectedCardState}
+          onClose={() => setSelectedCardId(undefined)}
           arrivalSize={selectedCardState.originalSize}
           formOption={formOption}
-          onClose={() => setSelectedCardId(undefined)}
           onCreateOrUpdateCardState={handleCreateOrUpdateCardState}
         />
       ) : (
@@ -538,33 +339,12 @@ export default function ItemRegisterContainer({ formOption }: TProps) {
               }}
             />
           </Box>
-          <Box>
-            <Fab
-              size="large"
-              sx={{
-                backgroundColor: "primary.main",
-                position: "absolute",
-                bottom: "12vh",
-                right: "12.5vw",
-              }}
-              onClick={() => setIsStockingIdInputDialogOpen(true)}
-            >
-              <QrCode2Icon fontSize="large" sx={{ color: "white" }} />
-            </Fab>
-            <StockingIdInputDialog
-              isOpen={isStockingIdInputDialogOpen}
-              onClose={() => setIsStockingIdInputDialogOpen(false)}
-              stockingDateCode={stockingDateCode}
-              onChangeStockingDateCode={(
-                e: ChangeEvent<HTMLInputElement | HTMLTextAreaElement>
-              ) => setStockingDateCode(e.target.value)}
-              stockingSequence={stockingSequence}
-              onChangeStockingSequence={(
-                e: ChangeEvent<HTMLInputElement | HTMLTextAreaElement>
-              ) => setStockingSequence(e.target.value)}
-              onClickAdd={handleClickAdd}
-            ></StockingIdInputDialog>
-          </Box>
+          <PreregisteredItemFetcher
+            onSetDuplicationTargetItemId={(itemId: number | undefined) =>
+              setDuplicationTargetItemId(itemId)
+            }
+            createNewCard={createNewCard}
+          />
         </>
       )}
     </>
